Refactor Select component structure
