+++ src/components/prompts/SelectShowcase.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { Select } from "@/components/ui";
+import { AnimatedSelect, NativeSelect, Select } from "@/components/ui";
 
 const items = [
   { value: "one", label: "One" },
@@ -10,8 +10,13 @@ export default function SelectShowcase() {
   const [value, setValue] = React.useState<string>();
   return (
     <div className="flex flex-col gap-4">
-      <Select items={items} value={value} onChange={setValue} placeholder="Animated" />
-      <Select variant="native" items={items} value={value} onChange={setValue} />
+      <AnimatedSelect
+        items={items}
+        value={value}
+        onChange={setValue}
+        placeholder="Animated"
+      />
+      <NativeSelect items={items} value={value} onChange={setValue} />
       <Select items={items} disabled placeholder="Disabled" />
     </div>
   );
