+++ src/components/ui/index.ts
@@ -63,6 +63,9 @@ export { default as SegmentedButton } from "./primitives/SegmentedButton";
 export * from "./primitives/SegmentedButton";
 export { default as Textarea } from "./primitives/Textarea";
 export * from "./primitives/Textarea";
+export { default as AnimatedSelect } from "./select/AnimatedSelect";
+export { default as NativeSelect } from "./select/NativeSelect";
+export * from "./select/shared";
 export { default as BackgroundPicker } from "./theme/BackgroundPicker";
 export * from "./theme/BackgroundPicker";
 export { default as ThemePicker } from "./theme/ThemePicker";
