+++ src/components/ui/select/Select.module.css
@@ -1,0 +1,238 @@
+.glitchTrigger {
+  --glitch-trigger: 1;
+}
+
+.glitchText {
+  --glitch-text: 1;
+}
+
+.caret {
+  transition:
+    transform var(--dur-quick) var(--ease-out),
+    filter var(--dur-quick) var(--ease-out);
+}
+
+.caretOpen {
+  transform: rotate(180deg);
+}
+
+.glitchTrigger:hover .caret {
+  animation: caret-jitter 0.9s steps(2, end) infinite;
+  filter: drop-shadow(0 0 4px hsl(var(--ring) / 0.55));
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .caret {
+    transition: none;
+  }
+  .glitchTrigger:hover .caret {
+    animation: none;
+    filter: none;
+  }
+}
+
+@keyframes caret-jitter {
+  0%,
+  100% {
+    transform: translateX(0) rotate(var(--rot, 0deg));
+  }
+  50% {
+    transform: translateX(0.6px) rotate(var(--rot, 0deg));
+  }
+}
+
+.gbIris,
+.gbChroma,
+.gbFlicker,
+.gbScan {
+  position: absolute;
+  inset: -1px;
+  border-radius: var(--radius-2xl);
+  pointer-events: none;
+}
+
+.gbIris {
+  padding: 1px;
+  background: conic-gradient(
+    from 180deg,
+    hsl(var(--ring) / 0),
+    hsl(var(--ring) / 0.55),
+    hsl(var(--accent-2) / 0.55),
+    hsl(var(--lav-deep) / 0.55),
+    hsl(var(--ring) / 0)
+  );
+  -webkit-mask:
+    linear-gradient(hsl(var(--foreground)) 0 0) content-box,
+    linear-gradient(hsl(var(--foreground)) 0 0);
+  -webkit-mask-composite: xor;
+  mask-composite: exclude;
+  opacity: 0.32;
+  animation: iris-rotate 10s linear infinite;
+}
+
+@keyframes iris-rotate {
+  to {
+    filter: hue-rotate(360deg);
+  }
+}
+
+.gbChroma {
+  padding: 1px;
+  background: conic-gradient(
+    from 90deg,
+    hsl(var(--ring) / 0),
+    hsl(var(--ring) / 0.7),
+    hsl(var(--accent-2) / 0.65),
+    hsl(var(--accent) / 0.65),
+    hsl(var(--ring) / 0)
+  );
+  -webkit-mask:
+    linear-gradient(hsl(var(--foreground)) 0 0) content-box,
+    linear-gradient(hsl(var(--foreground)) 0 0);
+  -webkit-mask-composite: xor;
+  mask-composite: exclude;
+  mix-blend-mode: screen;
+  opacity: 0.28;
+  animation: chroma-jitter 2.1s steps(6, end) infinite;
+}
+
+@keyframes chroma-jitter {
+  0%,
+  100% {
+    transform: translate(0, 0);
+  }
+  20% {
+    transform: translate(0.25px, -0.25px);
+  }
+  40% {
+    transform: translate(-0.25px, 0.25px);
+  }
+  60% {
+    transform: translate(0.15px, 0.15px);
+  }
+  80% {
+    transform: translate(-0.15px, -0.1px);
+  }
+}
+
+.gbFlicker {
+  inset: -2px;
+  border-radius: var(--radius-2xl);
+  background: radial-gradient(
+    120% 120% at 50% 50%,
+    hsl(var(--ring) / 0.18),
+    transparent 60%
+  );
+  filter: blur(7px) saturate(1.06);
+  mix-blend-mode: screen;
+  opacity: 0.18;
+  animation:
+    border-flicker 3.2s steps(24, end) infinite,
+    border-pulse 6s ease-in-out infinite alternate;
+}
+
+@keyframes border-flicker {
+  0%,
+  7%,
+  9%,
+  100% {
+    opacity: 0.16;
+  }
+  8% {
+    opacity: 0.46;
+  }
+  31% {
+    opacity: 0.22;
+  }
+  33% {
+    opacity: 0.4;
+  }
+  54% {
+    opacity: 0.2;
+  }
+  55% {
+    opacity: 0.46;
+  }
+  78% {
+    opacity: 0.24;
+  }
+}
+
+@keyframes border-pulse {
+  0%,
+  100% {
+    filter: blur(7px) saturate(1.02);
+  }
+  50% {
+    filter: blur(8px) saturate(1.18);
+  }
+}
+
+.gbScan {
+  padding: 1px;
+  -webkit-mask:
+    linear-gradient(hsl(var(--foreground)) 0 0) content-box,
+    linear-gradient(hsl(var(--foreground)) 0 0);
+  -webkit-mask-composite: xor;
+  mask-composite: exclude;
+  background: repeating-linear-gradient(
+    0deg,
+    hsl(var(--foreground) / 0.1) 0 1px,
+    transparent 2px 4px
+  );
+  mix-blend-mode: soft-light;
+  opacity: 0.2;
+  animation: scan-move 5.2s linear infinite;
+}
+
+@keyframes scan-move {
+  0% {
+    transform: translateY(-10%);
+  }
+  100% {
+    transform: translateY(10%);
+  }
+}
+
+.glitchTrigger[data-lit="true"] .gbIris,
+.glitchTrigger[data-open="true"] .gbIris {
+  opacity: 0.45;
+}
+
+.glitchTrigger[data-lit="true"] .gbChroma,
+.glitchTrigger[data-open="true"] .gbChroma {
+  opacity: 0.48;
+}
+
+.glitchTrigger[data-lit="true"] .gbFlicker,
+.glitchTrigger[data-open="true"] .gbFlicker {
+  opacity: 0.28;
+}
+
+[data-open="true"] .gbIris {
+  opacity: 0.38;
+}
+
+[data-open="true"] .gbChroma {
+  opacity: 0.42;
+}
+
+[data-open="true"] .gbFlicker {
+  opacity: 0.26;
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .gbIris,
+  .gbChroma,
+  .gbFlicker,
+  .gbScan,
+  .glitchTrigger:hover .caret {
+    animation: none;
+  }
+}
+
+.glitchText:hover {
+  text-shadow:
+    0.6px 0 hsl(var(--accent-2) / 0.45),
+    -0.6px 0 hsl(var(--lav-deep) / 0.45);
+}
